+++ src/ezdxf/addons/drawing/text.py
@@ -118,7 +118,7 @@ def _get_cap_height(text: AnyText) -> float:
 
 def _get_plain_text(text: AnyText) -> str:
     if isinstance(text, Text):
-        return remove_formatting(text.dxf.text)
+        return text.plain_text()
     elif isinstance(text, (MText, Attrib)):
         return remove_formatting(text.text)
     else: